errorMessageIdはIntで十分なのでIntにしました

+++ feature/communicate/src/main/java/com/cybozu/sample/kintone/spaces/feature/communicate/space/SpaceUiState.kt
@@ -5,5 +5,5 @@ import com.cybozu.sample.kintone.spaces.data.space.entity.Thread
 data class SpaceUiState(
     val threads: List<Thread> = emptyList(),
     val isLoading: Boolean = false,
-    val errorMessageId: Long? = null,
+    val errorMessageId: Int? = null,
 )
